add and move exception to own package
