+++ modules/core/src/main/java/net/ashald/envfile/exceptions/EnvSingleErrorException.java
@@ -1,0 +1,16 @@
+package net.ashald.envfile.exceptions;
+
+
+public class EnvSingleErrorException extends Exception {
+    public EnvSingleErrorException(String message) {
+        super(message);
+    }
+
+    public EnvSingleErrorException(Throwable cause) {
+        super(cause);
+    }
+
+    public EnvSingleErrorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
