+++ modules/core/src/main/java/net/ashald/envfile/providers/ip/IpEnvSingleProvider.java
@@ -1,7 +1,7 @@
 package net.ashald.envfile.providers.ip;
 
-import net.ashald.envfile.EnvFileErrorException;
 import net.ashald.envfile.EnvSingleProvider;
+import net.ashald.envfile.exceptions.EnvSingleErrorException;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
@@ -15,8 +15,9 @@ public class IpEnvSingleProvider implements EnvSingleProvider {
 
     }
 
+
     @Override
-    public @NotNull Map<String, String> process(@NotNull Map<String, String> runConfigEnv, String envVar, String selectedOption, @NotNull Map<String, String> aggregatedEnv) {
+    public @NotNull Map<String, String> process(@NotNull Map<String, String> runConfigEnv, String envVar, String selectedOption, @NotNull Map<String, String> aggregatedEnv) throws EnvSingleErrorException, IOException {
         return null;
     }
